+++ CHANGELOG.md
@@ -4,12 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## UNRELEASED
+## [0.14.0] 2017-09-12
 
 ### Added
 
 * Included `markdown-it-container` plugin by default
 * Included `markdown-it-attrs` plugin by default
+* New `html.yaml()` plugin to handle yaml files instead md
+
+### Changed
+
+* `html.permalink()` process all extensions, not only `.md`.
 
 ## [0.13.0] - 2017-08-10
 
@@ -167,6 +172,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 First version
 
+[0.14.0]: https://github.com/oscarotero/zume/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/oscarotero/zume/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/oscarotero/zume/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/oscarotero/zume/compare/v0.10.0...v0.11.0